requirements.txt
streamlit
agno
python-dotenv
yfinance
duckduckgo-search
requests
